requirements.txt
mysql-connector-python==8.0.30
rich==12.5.1
selenium==4.4.3
selenium_wire==4.6.5
